Service Unit Test Cleanup
